rename maze reporter class name
